+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.0.1-SNAPSHOT (Jan 5, 2016)
 - Support for HystrixRequestCaching in SinglePoolHttpTaskHandler
+- Bug fix: If service response object is plain String, an exception is thrown while decoding 
 
 ## 2.0.0  (Jan 4, 2016)
 - First public release